code/tenantstreamapp/requirements.txt
pandas
pyarrow
numpy
grpcio
grpcio-status
googleapis-common-protos
kafka-python
